move language name file from scripts folder to i18n folder
